In AC::TestCase#process, parse the request body to params as it is done in the ParamsParser middleware. fixes #13135.
this is another work-around to make AC::TestCase behave the way a real integration test works. as stated here https://github.com/rails/rails/issues/13135 we are planning to replace AC::TestCase logic with a faster IntegrationTest. instead of mocking all the environment variables, this will invoke the Rails stack and provide an environment that's identical to one in production.